simplify code
BC-1617
