--- apps/server/src/modules/user/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from './user.config';
