--- apps/server/src/modules/user/interfaces/user.config.ts
@@ -1,4 +1,0 @@
-export interface IUserConfig {
-	avaibleLanguage: string;
-	getAviableLanguages(): string[];
-}
