+++ apps/server/src/modules/user/user.module.ts
@@ -1,38 +1,14 @@
-import { Injectable, Module } from '@nestjs/common';
-import { Configuration } from '@hpi-schul-cloud/commons';
+import { Module } from '@nestjs/common';
 
 import { PermissionService } from '@shared/domain';
 import { UserRepo } from '@shared/repo';
 
 import { UserController } from './controller';
+import { UserConfig } from './user.config';
 import { UserUC } from './uc';
-import { IUserConfig } from './interfaces';
-
-@Injectable()
-export class UserConfig implements IUserConfig {
-	avaibleLanguage: string;
-
-	constructor() {
-		this.avaibleLanguage = Configuration.get('I18N__AVAILABLE_LANGUAGES') as string;
-	}
-
-	getAviableLanguages(): string[] {
-		const languages = this.avaibleLanguage.split(',');
-
-		return languages;
-	}
-}
 
 @Module({
 	controllers: [UserController],
-	providers: [
-		UserRepo,
-		PermissionService,
-		UserUC,
-		{
-			provide: 'User_Config',
-			useClass: UserConfig,
-		},
-	],
+	providers: [UserRepo, PermissionService, UserUC, UserConfig],
 })
 export class UserModule {}
